% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lynx.R
\name{lynx}
\alias{lynx}
\title{check links in a package}
\usage{
lynx(pkg = ".", files = c("DESCRIPTION", "NEWS.md", "README.md"))
}
\arguments{
\item{pkg}{path to a package}

\item{files}{(character) vector of files to look in. default:
DESCRIPTION, NEWS.md, README.md}
}
\description{
check links in a package
}
\details{
\itemize{
\item markdown files are converted to html first with
\code{markdown::markdownToHTML} to make it easier to extract URLs via
xpath with the \pkg{xml2} package
\item for non-markdown files we use regex to extract URLs
}
}
\examples{
\dontrun{
# within a package
# lynx(pkg="~/github/ropensci/rcrossref")
# lynx(pkg="~/github/ropensci/rnoaa")
}
}
