% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglynx-package.R
\docType{package}
\name{pkglynx-package}
\alias{pkglynx}
\alias{pkglynx-package}
\title{pkglynx: Check Package URLs}
\description{
Check Package URLs.
}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID})

}
\keyword{internal}
